@@ -1,58 +1,37 @@
-//  Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-// http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// 
+//   Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License. 
 
-//  Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-// http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// 
-
-using System;
 using System.Globalization;
-using System.IO;
 using System.Text;
-using Newtonsoft.Json;
 
-namespace RestSharp.Serializers.NewtonsoftJson
-{
-    public sealed class WriterBuffer : IDisposable
-    {
-        private readonly StringWriter _stringWriter;
-        private readonly JsonTextWriter _jsonTextWriter;
+namespace RestSharp.Serializers.NewtonsoftJson; 
 
-        public WriterBuffer(JsonSerializer jsonSerializer)
-        {
-            _stringWriter = new StringWriter(new StringBuilder(256), CultureInfo.InvariantCulture);
+sealed class WriterBuffer : IDisposable {
+    readonly StringWriter   _stringWriter;
+    readonly JsonTextWriter _jsonTextWriter;
 
-            _jsonTextWriter = new JsonTextWriter(_stringWriter)
-            {
-                Formatting = jsonSerializer.Formatting, CloseOutput = false
-            };
-        }
+    public WriterBuffer(JsonSerializer jsonSerializer) {
+        _stringWriter = new StringWriter(new StringBuilder(256), CultureInfo.InvariantCulture);
 
-        public JsonTextWriter GetJsonTextWriter() => _jsonTextWriter;
-        public StringWriter GetStringWriter() => _stringWriter;
-        public void Dispose() => _stringWriter.GetStringBuilder().Clear();
+        _jsonTextWriter = new JsonTextWriter(_stringWriter) {
+            Formatting = jsonSerializer.Formatting, CloseOutput = false
+        };
     }
+
+    public JsonTextWriter GetJsonTextWriter() => _jsonTextWriter;
+        
+    public StringWriter GetStringWriter() => _stringWriter;
+        
+    public void Dispose() => _stringWriter.GetStringBuilder().Clear();
 }
