@@ -1,24 +1,23 @@
-//   Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License. 
+//  Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+// http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-namespace RestSharp.Authenticators.OAuth
-{
-    public enum OAuthSignatureMethod { HmacSha1, HmacSha256, PlainText, RsaSha1 }
+namespace RestSharp.Authenticators.OAuth; 
+
+public enum OAuthSignatureMethod { HmacSha1, HmacSha256, PlainText, RsaSha1 }
     
-    public enum OAuthSignatureTreatment { Escaped, Unescaped }
+public enum OAuthSignatureTreatment { Escaped, Unescaped }
     
-    public enum OAuthParameterHandling { HttpAuthorizationHeader, UrlOrPostParameters }
+public enum OAuthParameterHandling { HttpAuthorizationHeader, UrlOrPostParameters }
     
-    public enum OAuthType { RequestToken, AccessToken, ProtectedResource, ClientAuthentication }
-}
+public enum OAuthType { RequestToken, AccessToken, ProtectedResource, ClientAuthentication }
