@@ -1,4 +1,4 @@
-﻿//   Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+﻿//   Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -12,32 +12,28 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License. 
 
-using System.Collections.Generic;
+namespace RestSharp.Authenticators.OAuth;
 
-namespace RestSharp.Authenticators.OAuth
-{
-    internal class WebPair
-    {
-        public WebPair(string name, string value)
-        {
-            Name  = name;
-            Value = value;
-        }
+class WebPair {
+    public WebPair(string name, string value, bool encode = false) {
+        Name     = name;
+        Value    = value;
+        WebValue = encode ? OAuthTools.UrlEncodeRelaxed(value) : value;
+        Encode   = encode;
+    }
 
-        public string Value { get; }
+    public string Name     { get; }
+    public string Value    { get; }
+    public string WebValue { get; }
+    public bool   Encode   { get; }
 
-        public string Name { get; }
-        
-        internal static WebPairComparer Comparer { get; } = new WebPairComparer();
+    internal static WebPairComparer Comparer { get; } = new();
 
-        internal class WebPairComparer : IComparer<WebPair>
-        {
-            public int Compare(WebPair x, WebPair y)
-            {
-                var compareName = string.CompareOrdinal(x?.Name, y?.Name);
+    internal class WebPairComparer : IComparer<WebPair> {
+        public int Compare(WebPair? x, WebPair? y) {
+            var compareName = string.CompareOrdinal(x?.Name, y?.Name);
 
-                return compareName != 0 ? compareName : string.CompareOrdinal(x?.Value, y?.Value);
-            }
+            return compareName != 0 ? compareName : string.CompareOrdinal(x?.Value, y?.Value);
         }
     }
 }
