@@ -1,232 +1,167 @@
-﻿//   Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License. 
+﻿//  Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+// http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-using System;
-using System.Net;
-using System.Threading;
-using System.Threading.Tasks;
-using RestSharp.Validation;
+using RestSharp.Extensions;
 
-namespace RestSharp
-{
-    public partial class RestClient
-    {
-        /// <summary>
-        /// Executes a GET-style request asynchronously, authenticating if needed
-        /// </summary>
-        /// <typeparam name="T">Target deserialization type</typeparam>
-        /// <param name="request">Request to be executed</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        public Task<IRestResponse<T>> ExecuteGetAsync<T>(IRestRequest request, CancellationToken cancellationToken = default)
-            => ExecuteAsync<T>(request, Method.GET, cancellationToken);
+namespace RestSharp;
 
-        /// <summary>
-        /// Executes a POST-style request asynchronously, authenticating if needed
-        /// </summary>
-        /// <typeparam name="T">Target deserialization type</typeparam>
-        /// <param name="request">Request to be executed</param>
-        /// <param name="cancellationToken">The cancellation token</param>
-        public Task<IRestResponse<T>> ExecutePostAsync<T>(IRestRequest request, CancellationToken cancellationToken = default)
-            => ExecuteAsync<T>(request, Method.POST, cancellationToken);
+public partial class RestClient {
+    /// <summary>
+    /// Executes the request asynchronously, authenticating if needed
+    /// </summary>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    public async Task<RestResponse> ExecuteAsync(RestRequest request, CancellationToken cancellationToken = default) {
+        var internalResponse = await ExecuteInternal(request, cancellationToken).ConfigureAwait(false);
 
-        /// <summary>
-        /// Executes a GET-style asynchronously, authenticating if needed
-        /// </summary>
-        /// <param name="request">Request to be executed</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        public Task<IRestResponse> ExecuteGetAsync(IRestRequest request, CancellationToken cancellationToken = default)
-            => ExecuteAsync(request, Method.GET, cancellationToken);
+        var response = new RestResponse();
 
-        /// <summary>
-        /// Executes a POST-style asynchronously, authenticating if needed
-        /// </summary>
-        /// <param name="request">Request to be executed</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        public Task<IRestResponse> ExecutePostAsync(IRestRequest request, CancellationToken cancellationToken = default)
-            => ExecuteAsync(request, Method.POST, cancellationToken);
-        
-        /// <summary>
-        /// Executes the request asynchronously, authenticating if needed
-        /// </summary>
-        /// <typeparam name="T">Target deserialization type</typeparam>
-        /// <param name="request">Request to be executed</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        public Task<IRestResponse<T>> ExecuteAsync<T>(IRestRequest request, CancellationToken cancellationToken = default)
-        {
-            if (request == null)
-                throw new ArgumentNullException(nameof(request));
-
-            var taskCompletionSource = new TaskCompletionSource<IRestResponse<T>>();
-
-            try
-            {
-                var async = ExecuteAsync<T>(
+        response = internalResponse.Exception == null
+            ? await RestResponse.FromHttpResponse(
+                    internalResponse.ResponseMessage!,
                     request,
-                    (response, _) =>
-                    {
-                        if (cancellationToken.IsCancellationRequested)
-                            taskCompletionSource.TrySetCanceled();
-                        // Don't run TrySetException, since we should set Error properties and swallow exceptions
-                        // to be consistent with sync methods
-                        else
-                            taskCompletionSource.TrySetResult(response);
-                    }
-                );
+                    Options.Encoding,
+                    CookieContainer.GetCookies(internalResponse.Url),
+                    CalculateResponseStatus,
+                    cancellationToken
+                )
+                .ConfigureAwait(false)
+            : AddError(response, internalResponse.Exception, internalResponse.TimeoutToken);
 
-                var registration =
-                    cancellationToken.Register(
-                        () =>
-                        {
-                            async.Abort();
-                            taskCompletionSource.TrySetCanceled();
-                        }
-                    );
+        response.Request = request;
+        response.Request.IncreaseNumAttempts();
 
-                taskCompletionSource.Task.ContinueWith(t => registration.Dispose(), cancellationToken);
-            }
-            catch (Exception ex)
-            {
-                taskCompletionSource.TrySetException(ex);
-            }
+        return Options.ThrowOnAnyError ? response.ThrowIfError() : response;
+    }
 
-            return taskCompletionSource.Task;
+    async Task<InternalResponse> ExecuteInternal(RestRequest request, CancellationToken cancellationToken) {
+        Ensure.NotNull(request, nameof(request));
+
+        using var requestContent = new RequestContent(this, request);
+
+        if (Authenticator != null)
+            await Authenticator.Authenticate(this, request).ConfigureAwait(false);
+
+        var httpMethod = AsHttpMethod(request.Method);
+        var url        = BuildUri(request);
+        var message    = new HttpRequestMessage(httpMethod, url) { Content = requestContent.BuildContent() };
+        message.Headers.Host         = Options.BaseHost;
+        message.Headers.CacheControl = Options.CachePolicy;
+
+        using var timeoutCts = new CancellationTokenSource(request.Timeout > 0 ? request.Timeout : int.MaxValue);
+        using var cts        = CancellationTokenSource.CreateLinkedTokenSource(timeoutCts.Token, cancellationToken);
+        var       ct         = cts.Token;
+
+        try {
+            var headers = new RequestHeaders()
+                .AddHeaders(request.Parameters)
+                .AddHeaders(DefaultParameters)
+                .AddAcceptHeader(AcceptedContentTypes);
+            message.AddHeaders(headers);
+
+            if (request.OnBeforeRequest != null)
+                await request.OnBeforeRequest(message).ConfigureAwait(false);
+
+            var responseMessage = await HttpClient.SendAsync(message, request.CompletionOption, ct).ConfigureAwait(false);
+
+            if (request.OnAfterRequest != null)
+                await request.OnAfterRequest(responseMessage).ConfigureAwait(false);
+
+            return new InternalResponse(responseMessage, url, null, timeoutCts.Token);
+        }
+        catch (Exception ex) {
+            return new InternalResponse(null, url, ex, timeoutCts.Token);
+        }
+    }
+
+    record InternalResponse(HttpResponseMessage? ResponseMessage, Uri Url, Exception? Exception, CancellationToken TimeoutToken);
+
+    /// <summary>
+    /// A specialized method to download files as streams.
+    /// </summary>
+    /// <param name="request">Pre-configured request instance.</param>
+    /// <param name="cancellationToken"></param>
+    /// <returns>The downloaded stream.</returns>
+    [PublicAPI]
+    public async Task<Stream?> DownloadStreamAsync(RestRequest request, CancellationToken cancellationToken = default) {
+        // Make sure we only read the headers so we can stream the content body efficiently
+        request.CompletionOption = HttpCompletionOption.ResponseHeadersRead;
+        var response = await ExecuteInternal(request, cancellationToken).ConfigureAwait(false);
+
+        if (response.Exception != null) {
+            return Options.ThrowOnAnyError ? throw response.Exception : null;
         }
 
-        /// <summary>
-        /// Executes the request asynchronously, authenticating if needed
-        /// </summary>
-        /// <param name="request">Request to be executed</param>
-        /// <param name="httpMethod">Override the request method</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        public Task<IRestResponse> ExecuteAsync(
-            IRestRequest request,
-            Method httpMethod,
-            CancellationToken cancellationToken = default
-        )
-        {
-            Ensure.NotNull(request, nameof(request));
+        if (response.ResponseMessage == null) return null;
 
-            request.Method = httpMethod;
-            return ExecuteAsync(request, cancellationToken);
+        if (request.ResponseWriter != null) {
+#if NETSTANDARD
+            using var stream = await response.ResponseMessage.ReadResponse(cancellationToken).ConfigureAwait(false);
+#else
+            await using var stream = await response.ResponseMessage.ReadResponse(cancellationToken).ConfigureAwait(false);
+#endif
+            return request.ResponseWriter(stream!);
         }
 
-        /// <summary>
-        /// Executes the request asynchronously, authenticating if needed
-        /// </summary>
-        /// <typeparam name="T">Target deserialization type</typeparam>
-        /// <param name="request">Request to be executed</param>
-        /// <param name="httpMethod">Override the request method</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        public Task<IRestResponse<T>> ExecuteAsync<T>(
-            IRestRequest request,
-            Method httpMethod,
-            CancellationToken cancellationToken = default
-        )
-        {
-            Ensure.NotNull(request, nameof(request));
+        return await response.ResponseMessage.ReadResponse(cancellationToken).ConfigureAwait(false);
+    }
 
-            request.Method = httpMethod;
-            return ExecuteAsync<T>(request, cancellationToken);
-        }
+    static RestResponse AddError(RestResponse response, Exception exception, CancellationToken timeoutToken) {
+        response.ResponseStatus = exception is OperationCanceledException
+            ? TimedOut() ? ResponseStatus.TimedOut : ResponseStatus.Aborted
+            : ResponseStatus.Error;
 
-        /// <inheritdoc />
-        public Task<IRestResponse> ExecuteAsync(IRestRequest request, CancellationToken token = default)
-        {
-            Ensure.NotNull(request, nameof(request));
+        response.ErrorMessage   = exception.Message;
+        response.ErrorException = exception;
 
-            var taskCompletionSource = new TaskCompletionSource<IRestResponse>();
+        return response;
 
-            try
-            {
-                var async = ExecuteAsync(
-                    request,
-                    (response, _) =>
-                    {
-                        if (token.IsCancellationRequested)
-                            taskCompletionSource.TrySetCanceled();
-                        // Don't run TrySetException, since we should set Error
-                        // properties and swallow exceptions to be consistent
-                        // with sync methods
-                        else
-                            taskCompletionSource.TrySetResult(response);
-                    }
-                );
+        bool TimedOut() => timeoutToken.IsCancellationRequested || exception.Message.Contains("HttpClient.Timeout");
+    }
 
-                var registration =
-                    token.Register(
-                        () =>
-                        {
-                            async.Abort();
-                            taskCompletionSource.TrySetCanceled();
-                        }
-                    );
+    static HttpMethod AsHttpMethod(Method method)
+        => method switch {
+            Method.Get     => HttpMethod.Get,
+            Method.Post    => HttpMethod.Post,
+            Method.Put     => HttpMethod.Put,
+            Method.Delete  => HttpMethod.Delete,
+            Method.Head    => HttpMethod.Head,
+            Method.Options => HttpMethod.Options,
+#if NETSTANDARD
+            Method.Patch => new HttpMethod("PATCH"),
+#else
+            Method.Patch   => HttpMethod.Patch,
+#endif
+            Method.Merge  => new HttpMethod("MERGE"),
+            Method.Copy   => new HttpMethod("COPY"),
+            Method.Search => new HttpMethod("SEARCH"),
+            _             => throw new ArgumentOutOfRangeException()
+        };
+}
 
-                taskCompletionSource.Task.ContinueWith(t => registration.Dispose(), token);
-            }
-            catch (Exception ex)
-            {
-                taskCompletionSource.TrySetException(ex);
-            }
+public static class ResponseThrowExtension {
+    public static RestResponse ThrowIfError(this RestResponse response) {
+        var exception = response.GetException();
+        if (exception != null) throw exception;
 
-            return taskCompletionSource.Task;
-        }
+        return response;
+    }
+    
+    public static RestResponse<T> ThrowIfError<T>(this RestResponse<T> response) {
+        var exception = response.GetException();
+        if (exception != null) throw exception;
 
-        RestRequestAsyncHandle ExecuteAsync(
-            IRestRequest request,
-            Action<IRestResponse, RestRequestAsyncHandle> callback,
-            string httpMethod,
-            Func<IHttp, Action<HttpResponse>, string, HttpWebRequest> getWebRequest
-        )
-        {
-            request.SerializeRequestBody(Serializers, request.XmlSerializer, request.JsonSerializer);
-
-            AuthenticateIfNeeded(request);
-
-            var http = ConfigureHttp(request);
-
-            request.OnBeforeRequest?.Invoke(http);
-
-            var asyncHandle = new RestRequestAsyncHandle();
-
-            Action<HttpResponse> responseCb = ProcessResponse;
-
-            if (UseSynchronizationContext && SynchronizationContext.Current != null)
-            {
-                var ctx = SynchronizationContext.Current;
-                var cb  = responseCb;
-
-                responseCb = resp => ctx.Post(s => cb(resp), null);
-            }
-
-            asyncHandle.WebRequest = getWebRequest(http, responseCb, httpMethod);
-
-            return asyncHandle;
-
-            void ProcessResponse(IHttpResponse httpResponse)
-            {
-                var restResponse = RestResponse.FromHttpResponse(httpResponse, request);
-                restResponse.Request.IncreaseNumAttempts();
-                callback(restResponse, asyncHandle);
-            }
-        }
-
-        void DeserializeResponse<T>(
-            IRestRequest request,
-            Action<IRestResponse<T>, RestRequestAsyncHandle> callback,
-            IRestResponse response,
-            RestRequestAsyncHandle asyncHandle
-        )
-            => callback(Deserialize<T>(request, response), asyncHandle);
+        return response;
     }
 }
