@@ -1,641 +1,238 @@
-﻿//   Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License. 
+﻿//  Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+// http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net;
-using System.Net.Cache;
-using System.Net.Security;
-using System.Reflection;
-using System.Security.Cryptography.X509Certificates;
+using System.Net.Http.Headers;
 using System.Text;
-using System.Text.RegularExpressions;
-using JetBrains.Annotations;
 using RestSharp.Authenticators;
-using RestSharp.Authenticators.OAuth.Extensions;
-using RestSharp.Deserializers;
 using RestSharp.Extensions;
-using RestSharp.Serialization;
-using RestSharp.Serialization.Json;
-using RestSharp.Serialization.Xml;
-using static System.String;
 
-namespace RestSharp
-{
+// ReSharper disable VirtualMemberCallInConstructor
+#pragma warning disable 618
+
+namespace RestSharp;
+
+/// <summary>
+/// Client to translate RestRequests into Http requests and process response result
+/// </summary>
+public partial class RestClient : IDisposable {
+    public CookieContainer CookieContainer { get; }
+
     /// <summary>
-    /// Client to translate RestRequests into Http requests and process response result
+    /// Content types that will be sent in the Accept header. The list is populated from the known serializers.
+    /// If you need to send something else by default, set this property to a different value.
     /// </summary>
-    [PublicAPI]
-    public partial class RestClient : IRestClient
-    {
-        static readonly Version Version = new AssemblyName(typeof(RestClient).Assembly.FullName).Version;
+    public string[] AcceptedContentTypes { get; set; } = null!;
 
-        static readonly Regex StructuredSyntaxSuffixRegex = new Regex(@"\+\w+$");
+    /// <summary>
+    /// Function to calculate the response status. By default, the status will be Completed if it was successful, or NotFound.
+    /// </summary>
+    public CalculateResponseStatus CalculateResponseStatus { get; set; } = httpResponse
+        => httpResponse.IsSuccessStatusCode || httpResponse.StatusCode == HttpStatusCode.NotFound
+            ? ResponseStatus.Completed
+            : ResponseStatus.Error;
 
-        static readonly Regex StructuredSyntaxSuffixWildcardRegex = new Regex(@"^\*\+\w+$");
+    HttpClient HttpClient { get; }
 
-        static readonly ParameterType[] MultiParameterTypes =
-            {ParameterType.QueryString, ParameterType.GetOrPost};
+    public RestClientOptions Options { get; }
 
-        /// <summary>
-        /// Default constructor that registers default content handlers
-        /// </summary>
-        public RestClient()
-        {
-            Encoding               = Encoding.UTF8;
-            ContentHandlers        = new Dictionary<string, Func<IDeserializer>>();
-            Serializers            = new Dictionary<DataFormat, IRestSerializer>();
-            AcceptTypes            = new List<string>();
-            DefaultParameters      = new List<Parameter>();
-            AutomaticDecompression = true;
+    public RestClient(RestClientOptions options, Action<HttpRequestHeaders>? configureDefaultHeaders = null) {
+        UseDefaultSerializers();
 
-            // register default serializers
-            UseSerializer<JsonSerializer>();
-            UseSerializer<XmlRestSerializer>();
+        Options            = options;
+        CookieContainer    = Options.CookieContainer ?? new CookieContainer();
+        _disposeHttpClient = true;
 
-            FollowRedirects = true;
+        var handler = new HttpClientHandler();
+        ConfigureHttpMessageHandler(handler);
+
+        var finalHandler = Options.ConfigureMessageHandler?.Invoke(handler) ?? handler;
+
+        HttpClient = new HttpClient(finalHandler);
+        ConfigureHttpClient(HttpClient);
+        configureDefaultHeaders?.Invoke(HttpClient.DefaultRequestHeaders);
+    }
+
+    /// <summary>
+    /// Creates an instance of RestClient using the default <see cref="RestClientOptions"/>
+    /// </summary>
+    public RestClient() : this(new RestClientOptions()) { }
+
+    /// <inheritdoc />
+    /// <summary>
+    /// Sets the BaseUrl property for requests made by this client instance
+    /// </summary>
+    /// <param name="baseUrl"></param>
+    public RestClient(Uri baseUrl) : this(new RestClientOptions { BaseUrl = baseUrl }) { }
+
+    /// <inheritdoc />
+    /// <summary>
+    /// Sets the BaseUrl property for requests made by this client instance
+    /// </summary>
+    /// <param name="baseUrl"></param>
+    public RestClient(string baseUrl) : this(new Uri(Ensure.NotEmptyString(baseUrl, nameof(baseUrl)))) { }
+
+    public RestClient(HttpClient httpClient, bool disposeHttpClient = false) {
+        UseDefaultSerializers();
+
+        HttpClient         = httpClient;
+        CookieContainer    = new CookieContainer();
+        Options            = new RestClientOptions();
+        _disposeHttpClient = disposeHttpClient;
+
+        if (httpClient.BaseAddress != null) {
+            Options.BaseUrl = httpClient.BaseAddress;
+        }
+    }
+
+    public RestClient(HttpClient httpClient, RestClientOptions options, bool disposeHttpClient = false) {
+        if (options.CookieContainer != null) {
+            throw new ArgumentException("Custom cookie container cannot be added to the HttpClient instance", nameof(options.CookieContainer));
         }
 
-        /// <inheritdoc />
-        /// <summary>
-        /// Sets the BaseUrl property for requests made by this client instance
-        /// </summary>
-        /// <param name="baseUrl"></param>
-        public RestClient(Uri baseUrl) : this() => BaseUrl = baseUrl;
+        UseDefaultSerializers();
 
-        /// <inheritdoc />
-        /// <summary>
-        /// Sets the BaseUrl property for requests made by this client instance
-        /// </summary>
-        /// <param name="baseUrl"></param>
-        public RestClient(string baseUrl) : this()
-        {
-            if (baseUrl.IsEmpty())
-                throw new ArgumentNullException(nameof(baseUrl));
+        HttpClient         = httpClient;
+        CookieContainer    = new CookieContainer();
+        Options            = options;
+        _disposeHttpClient = disposeHttpClient;
 
-            BaseUrl = new Uri(baseUrl);
+        if (httpClient.BaseAddress != null && options.BaseUrl == null) {
+            Options.BaseUrl = httpClient.BaseAddress;
         }
 
-        IDictionary<string, Func<IDeserializer>> ContentHandlers { get; }
-        internal IDictionary<DataFormat, IRestSerializer> Serializers { get; }
-        Func<string, string> Encode { get; set; } = s => s.UrlEncode();
-        Func<string, Encoding, string> EncodeQuery { get; set; } = (s, encoding) => s.UrlEncode(encoding);
-        IList<string> AcceptTypes { get; }
-        Action<HttpWebRequest> WebRequestConfigurator { get; set; }
+        ConfigureHttpClient(HttpClient);
+    }
 
-        /// <inheritdoc />
-        [Obsolete("Use the overload that accepts the delegate factory")]
-        public IRestClient UseSerializer(IRestSerializer serializer) => this.With(x => x.UseSerializer(() => serializer));
+    /// <summary>
+    /// Creates a new instance of RestSharp using the message handler provided. By default, HttpClient disposes the provided handler
+    /// when the client itself is disposed. If you want to keep the handler not disposed, set disposeHandler argument to false.
+    /// </summary>
+    /// <param name="handler">Message handler instance to use for HttpClient</param>
+    /// <param name="disposeHandler">Dispose the handler when disposing RestClient, true by default</param>
+    public RestClient(HttpMessageHandler handler, bool disposeHandler = true) : this(new HttpClient(handler, disposeHandler), true) { }
 
-        /// <inheritdoc />
-        public IRestClient UseUrlEncoder(Func<string, string> encoder) => this.With(x => x.Encode = encoder);
+    void ConfigureHttpClient(HttpClient httpClient) {
+        if (Options.MaxTimeout > 0) httpClient.Timeout = TimeSpan.FromMilliseconds(Options.MaxTimeout);
+        if (httpClient.DefaultRequestHeaders.UserAgent.All(x => x.Product.Name != "RestSharp")) {
+            httpClient.DefaultRequestHeaders.UserAgent.ParseAdd(Options.UserAgent);
+        }
+        if (Options.Expect100Continue != null)
+            httpClient.DefaultRequestHeaders.ExpectContinue = Options.Expect100Continue;
+    }
 
-        /// <inheritdoc />
-        public IRestClient UseQueryEncoder(Func<string, Encoding, string> queryEncoder) => this.With(x => x.EncodeQuery = queryEncoder);
+    void ConfigureHttpMessageHandler(HttpClientHandler handler) {
+        handler.Credentials            = Options.Credentials;
+        handler.UseDefaultCredentials  = Options.UseDefaultCredentials;
+        handler.CookieContainer        = CookieContainer;
+        handler.AutomaticDecompression = Options.AutomaticDecompression;
+        handler.PreAuthenticate        = Options.PreAuthenticate;
+        handler.AllowAutoRedirect      = Options.FollowRedirects;
+        handler.Proxy                  = Options.Proxy;
 
-        /// <inheritdoc />
-        public bool AutomaticDecompression { get; set; }
+        if (Options.RemoteCertificateValidationCallback != null)
+            handler.ServerCertificateCustomValidationCallback =
+                (request, cert, chain, errors) => Options.RemoteCertificateValidationCallback(request, cert, chain, errors);
 
-        /// <inheritdoc />
-        public int? MaxRedirects { get; set; }
-
-        /// <inheritdoc />
-        public X509CertificateCollection ClientCertificates { get; set; }
-
-        /// <inheritdoc />
-        public IWebProxy Proxy { get; set; }
-
-        /// <inheritdoc />
-        public RequestCachePolicy CachePolicy { get; set; }
-
-        /// <inheritdoc />
-        public bool Pipelined { get; set; }
-
-        /// <inheritdoc />
-        public bool FollowRedirects { get; set; }
-
-        /// <inheritdoc />
-        public CookieContainer CookieContainer { get; set; }
-
-        /// <inheritdoc />
-        public string UserAgent { get; set; }
-
-        /// <inheritdoc />
-        public int Timeout { get; set; }
-
-        /// <inheritdoc />
-        public int ReadWriteTimeout { get; set; }
-
-        /// <inheritdoc />
-        public bool UseSynchronizationContext { get; set; }
-
-        /// <inheritdoc />
-        public IAuthenticator Authenticator { get; set; }
-
-        /// <inheritdoc />
-        public virtual Uri BaseUrl { get; set; }
-
-        /// <inheritdoc />
-        public Encoding Encoding { get; set; }
-
-        /// <inheritdoc />
-        public bool PreAuthenticate { get; set; }
-
-        /// <inheritdoc />
-        public bool ThrowOnDeserializationError { get; set; } = false;
-
-        /// <inheritdoc />
-        public bool FailOnDeserializationError { get; set; } = true;
-
-        /// <inheritdoc />
-        public bool ThrowOnAnyError { get; set; } = false;
-
-        /// <inheritdoc />
-        public bool UnsafeAuthenticatedConnectionSharing { get; set; }
-
-        /// <inheritdoc />
-        public string ConnectionGroupName { get; set; }
-
-        /// <inheritdoc />
-        public RemoteCertificateValidationCallback RemoteCertificateValidationCallback { get; set; }
-
-        /// <inheritdoc />
-        public IList<Parameter> DefaultParameters { get; }
-
-        /// <inheritdoc />>
-        public string BaseHost { get; set; }
-
-        /// <inheritdoc />>
-        public bool AllowMultipleDefaultParametersWithSameName { get; set; } = false;
-
-        /// <inheritdoc />>
-        public void AddHandler(string contentType, Func<IDeserializer> deserializerFactory)
-        {
-            ContentHandlers[contentType] = deserializerFactory;
-
-            if (contentType == "*" || IsWildcardStructuredSuffixSyntax(contentType)) return;
-
-            if (!AcceptTypes.Contains(contentType))
-                AcceptTypes.Add(contentType);
-
-            // add Accept header based on registered deserializers
-            var accepts = AcceptTypes.JoinToString(", ");
-
-            this.AddOrUpdateDefaultParameter(new Parameter("Accept", accepts, ParameterType.HttpHeader));
+        if (Options.ClientCertificates != null) {
+            handler.ClientCertificates.AddRange(Options.ClientCertificates);
+            handler.ClientCertificateOptions = ClientCertificateOption.Manual;
         }
 
-        /// <inheritdoc />>
-        [Obsolete("Use the overload that accepts a factory delegate")]
-        public void AddHandler(string contentType, IDeserializer deserializer) => AddHandler(contentType, () => deserializer);
+        if (Options.MaxRedirects.HasValue) handler.MaxAutomaticRedirections = Options.MaxRedirects.Value;
+    }
 
-        /// <inheritdoc />
-        public void RemoveHandler(string contentType)
-        {
-            ContentHandlers.Remove(contentType);
-            AcceptTypes.Remove(contentType);
-            this.RemoveDefaultParameter("Accept");
+    internal Func<string, string> Encode { get; set; } = s => s.UrlEncode();
+
+    internal Func<string, Encoding, string> EncodeQuery { get; set; } = (s, encoding) => s.UrlEncode(encoding)!;
+
+    /// <summary>
+    /// Authenticator that will be used to populate request with necessary authentication data
+    /// </summary>
+    public IAuthenticator? Authenticator { get; set; }
+
+    public ParametersCollection DefaultParameters { get; } = new();
+
+    /// <summary>
+    /// Add a parameter to use on every request made with this client instance
+    /// </summary>
+    /// <param name="parameter">Parameter to add</param>
+    /// <returns></returns>
+    public RestClient AddDefaultParameter(Parameter parameter) {
+        if (parameter.Type == ParameterType.RequestBody)
+            throw new NotSupportedException(
+                "Cannot set request body using default parameters. Use Request.AddBody() instead."
+            );
+
+        if (!Options.AllowMultipleDefaultParametersWithSameName &&
+            !MultiParameterTypes.Contains(parameter.Type) &&
+            DefaultParameters.Any(x => x.Name == parameter.Name)) {
+            throw new ArgumentException("A default parameters with the same name has already been added", nameof(parameter));
         }
 
-        /// <inheritdoc />
-        public void ClearHandlers()
-        {
-            ContentHandlers.Clear();
-            AcceptTypes.Clear();
-            this.RemoveDefaultParameter("Accept");
+        DefaultParameters.AddParameter(parameter);
+
+        return this;
+    }
+
+    static readonly ParameterType[] MultiParameterTypes = { ParameterType.QueryString, ParameterType.GetOrPost };
+
+    public Uri BuildUri(RestRequest request) {
+        DoBuildUriValidations(request);
+
+        var (uri, resource) = Options.BaseUrl.GetUrlSegmentParamsValues(request.Resource, Encode, request.Parameters, DefaultParameters);
+        var mergedUri = uri.MergeBaseUrlAndResource(resource);
+
+        var finalUri = mergedUri.ApplyQueryStringParamsValuesToUri(
+            request.Method,
+            Options.Encoding,
+            EncodeQuery,
+            request.Parameters,
+            DefaultParameters
+        );
+        return finalUri;
+    }
+
+    internal Uri BuildUriWithoutQueryParameters(RestRequest request) {
+        DoBuildUriValidations(request);
+        var (uri, resource) = Options.BaseUrl.GetUrlSegmentParamsValues(request.Resource, Encode, request.Parameters, DefaultParameters);
+        return uri.MergeBaseUrlAndResource(resource);
+    }
+
+    internal void AssignAcceptedContentTypes()
+        => AcceptedContentTypes = Serializers.SelectMany(x => x.Value.AcceptedContentTypes).Distinct().ToArray();
+
+    void DoBuildUriValidations(RestRequest request) {
+        if (Options.BaseUrl == null && !request.Resource.ToLowerInvariant().StartsWith("http"))
+            throw new ArgumentOutOfRangeException(
+                nameof(request),
+                "Request resource doesn't contain a valid scheme for an empty base URL of the client"
+            );
+    }
+
+    readonly bool _disposeHttpClient;
+    bool          _disposed;
+
+    protected virtual void Dispose(bool disposing) {
+        if (disposing && !_disposed) {
+            _disposed = true;
+            if (_disposeHttpClient) HttpClient.Dispose();
         }
-
-        /// <inheritdoc />
-        public IRestResponse<T> Deserialize<T>(IRestResponse response) => Deserialize<T>(response.Request, response);
-
-        /// <inheritdoc />
-        public void ConfigureWebRequest(Action<HttpWebRequest> configurator) => WebRequestConfigurator = configurator;
-
-        /// <inheritdoc />
-        public Uri BuildUri(IRestRequest request)
-        {
-            DoBuildUriValidations(request);
-
-            var applied = GetUrlSegmentParamsValues(request);
-
-            var mergedUri = MergeBaseUrlAndResource(applied.Uri, applied.Resource);
-
-            var finalUri = ApplyQueryStringParamsValuesToUri(mergedUri, request);
-
-            return new Uri(finalUri);
-        }
-
-        string IRestClient.BuildUriWithoutQueryParameters(IRestRequest request)
-        {
-            DoBuildUriValidations(request);
-
-            var applied = GetUrlSegmentParamsValues(request);
-
-            return MergeBaseUrlAndResource(applied.Uri, applied.Resource);
-        }
-
-        /// <inheritdoc />
-        public IRestClient UseSerializer(Func<IRestSerializer> serializerFactory)
-        {
-            var instance = serializerFactory();
-            Serializers[instance.DataFormat] = instance;
-
-            AddHandler(serializerFactory, instance.SupportedContentTypes);
-
-            return this;
-        }
-
-        /// <inheritdoc />
-        public IRestClient UseSerializer<T>() where T : IRestSerializer, new() => UseSerializer(() => new T());
-
-        void AddHandler(Func<IDeserializer> deserializerFactory, params string[] contentTypes)
-        {
-            foreach (var contentType in contentTypes)
-                AddHandler(contentType, deserializerFactory);
-        }
-
-        void DoBuildUriValidations(IRestRequest request)
-        {
-            if (BaseUrl == null && !request.Resource.ToLower().StartsWith("http"))
-                throw new ArgumentOutOfRangeException(
-                    nameof(request),
-                    "Request resource doesn't contain a valid scheme for an empty client base URL"
-                );
-
-            var nullValuedParams = request.Parameters
-                .Where(p => p.Type == ParameterType.UrlSegment && p.Value == null)
-                .Select(p => p.Name)
-                .ToArray();
-
-            if (nullValuedParams.Any())
-            {
-                var names = nullValuedParams.JoinToString(", ", name => $"'{name}'");
-
-                throw new ArgumentException(
-                    $"Cannot build uri when url segment parameter(s) {names} value is null.",
-                    nameof(request)
-                );
-            }
-        }
-
-        UrlSegmentParamsValues GetUrlSegmentParamsValues(IRestRequest request)
-        {
-            var assembled = BaseUrl == null ? "" : request.Resource;
-            var baseUrl   = BaseUrl ?? new Uri(request.Resource);
-
-            var hasResource = !assembled.IsEmpty();
-            var parameters  = request.Parameters.Where(p => p.Type == ParameterType.UrlSegment).ToList();
-            parameters.AddRange(DefaultParameters.Where(p => p.Type == ParameterType.UrlSegment));
-            var builder = new UriBuilder(baseUrl);
-
-            foreach (var parameter in parameters)
-            {
-                var paramPlaceHolder = $"{{{parameter.Name}}}";
-                var paramValue       = Encode(parameter.Value.ToString());
-
-                if (hasResource)
-                    assembled = assembled.Replace(paramPlaceHolder, paramValue);
-
-                builder.Path = builder.Path.UrlDecode().Replace(paramPlaceHolder, paramValue);
-            }
-
-            return new UrlSegmentParamsValues(builder.Uri, assembled);
-        }
-
-        static string MergeBaseUrlAndResource(Uri baseUrl, string resource)
-        {
-            var assembled = resource;
-
-            if (!IsNullOrEmpty(assembled) && assembled.StartsWith("/")) assembled = assembled.Substring(1);
-
-            if (baseUrl == null || IsNullOrEmpty(baseUrl.AbsoluteUri)) return assembled;
-
-            var usingBaseUri                                                                  = baseUrl;
-            if (!baseUrl.AbsoluteUri.EndsWith("/") && !IsNullOrEmpty(assembled)) usingBaseUri = new Uri(baseUrl.AbsoluteUri + "/");
-
-            return assembled != null ? new Uri(usingBaseUri, assembled).AbsoluteUri : baseUrl.AbsoluteUri;
-        }
-
-        string ApplyQueryStringParamsValuesToUri(string mergedUri, IRestRequest request)
-        {
-            var parameters = GetQueryStringParameters(request).ToList();
-            parameters.AddRange(GetDefaultQueryStringParameters(request));
-
-            if (!parameters.Any()) return mergedUri;
-
-            var separator = mergedUri != null && mergedUri.Contains("?") ? "&" : "?";
-
-            return Concat(mergedUri, separator, EncodeParameters(parameters, Encoding));
-        }
-
-        IEnumerable<Parameter> GetDefaultQueryStringParameters(IRestRequest request)
-            => request.Method != Method.POST && request.Method != Method.PUT && request.Method != Method.PATCH
-                ? DefaultParameters
-                    .Where(
-                        p => p.Type == ParameterType.GetOrPost   ||
-                            p.Type  == ParameterType.QueryString ||
-                            p.Type  == ParameterType.QueryStringWithoutEncode
-                    )
-                : DefaultParameters
-                    .Where(
-                        p => p.Type == ParameterType.QueryString ||
-                            p.Type  == ParameterType.QueryStringWithoutEncode
-                    );
-
-        static IEnumerable<Parameter> GetQueryStringParameters(IRestRequest request)
-            => request.Method != Method.POST && request.Method != Method.PUT && request.Method != Method.PATCH
-                ? request.Parameters
-                    .Where(
-                        p => p.Type == ParameterType.GetOrPost   ||
-                            p.Type  == ParameterType.QueryString ||
-                            p.Type  == ParameterType.QueryStringWithoutEncode
-                    )
-                : request.Parameters
-                    .Where(
-                        p => p.Type == ParameterType.QueryString ||
-                            p.Type  == ParameterType.QueryStringWithoutEncode
-                    );
-
-        Func<IDeserializer> GetHandler(string contentType)
-        {
-            if (contentType.IsEmpty() && ContentHandlers.ContainsKey("*"))
-                return ContentHandlers["*"];
-
-            if (contentType.IsEmpty())
-                return ContentHandlers.First().Value;
-
-            var semicolonIndex = contentType.IndexOf(';');
-
-            if (semicolonIndex > -1)
-                contentType = contentType.Substring(0, semicolonIndex);
-
-            if (ContentHandlers.TryGetValue(contentType, out var contentHandler))
-                return contentHandler;
-
-            // Avoid unnecessary use of regular expressions in checking for structured syntax suffix by looking for a '+' first
-            if (contentType.IndexOf('+') >= 0)
-            {
-                // https://tools.ietf.org/html/rfc6839#page-4
-                var structuredSyntaxSuffixMatch = StructuredSyntaxSuffixRegex.Match(contentType);
-
-                if (structuredSyntaxSuffixMatch.Success)
-                {
-                    var structuredSyntaxSuffixWildcard = "*" + structuredSyntaxSuffixMatch.Value;
-                    if (ContentHandlers.TryGetValue(structuredSyntaxSuffixWildcard, out var contentHandlerWildcard)) return contentHandlerWildcard;
-                }
-            }
-
-            return ContentHandlers.ContainsKey("*") ? ContentHandlers["*"] : null;
-        }
-
-        void AuthenticateIfNeeded(IRestRequest request) => Authenticator?.Authenticate(this, request);
-
-        string EncodeParameters(IEnumerable<Parameter> parameters, Encoding encoding)
-            => Join("&", parameters.Select(parameter => EncodeParameter(parameter, encoding)).ToArray());
-
-        string EncodeParameter(Parameter parameter, Encoding encoding)
-        {
-            return
-                parameter.Type == ParameterType.QueryStringWithoutEncode
-                    ? $"{parameter.Name}={StringOrEmpty(parameter.Value)}"
-                    : $"{EncodeQuery(parameter.Name, encoding)}={EncodeQuery(StringOrEmpty(parameter.Value), encoding)}";
-
-            static string StringOrEmpty(object value) => value == null ? "" : value.ToString();
-        }
-
-        IHttp ConfigureHttp(IRestRequest request)
-        {
-            var http = new Http
-            {
-                Encoding                = Encoding,
-                AlwaysMultipartFormData = request.AlwaysMultipartFormData,
-                UseDefaultCredentials   = request.UseDefaultCredentials,
-                ResponseWriter          = request.ResponseWriter,
-                AdvancedResponseWriter  = request.AdvancedResponseWriter,
-                CookieContainer         = CookieContainer,
-                AutomaticDecompression  = AutomaticDecompression,
-                WebRequestConfigurator  = WebRequestConfigurator,
-                Encode                  = Encode
-            };
-
-            var requestParameters = new List<Parameter>();
-            requestParameters.AddRange(request.Parameters);
-
-            // move RestClient.DefaultParameters into Request.Parameters
-            foreach (var defaultParameter in DefaultParameters)
-            {
-                var parameterExists =
-                    request.Parameters.Any(
-                        p =>
-                            p.Name != null
-                            && p.Name.Equals(defaultParameter.Name, StringComparison.InvariantCultureIgnoreCase)
-                            && p.Type == defaultParameter.Type
-                    );
-
-                if (AllowMultipleDefaultParametersWithSameName)
-                {
-                    var isMultiParameter = MultiParameterTypes.Any(pt => pt == defaultParameter.Type);
-                    parameterExists = !isMultiParameter && parameterExists;
-                }
-
-                if (!parameterExists) requestParameters.Add(defaultParameter);
-            }
-
-            // Add Accept header based on registered deserializers if none has been set by the caller.
-            if (requestParameters.All(
-                p => !p.Name.EqualsIgnoreCase("accept")
-            ))
-            {
-                var accepts = Join(", ", AcceptTypes);
-
-                requestParameters.Add(new Parameter("Accept", accepts, ParameterType.HttpHeader));
-            }
-
-            http.Url                                  = BuildUri(request);
-            http.Host                                 = BaseHost;
-            http.PreAuthenticate                      = PreAuthenticate;
-            http.UnsafeAuthenticatedConnectionSharing = UnsafeAuthenticatedConnectionSharing;
-
-            var userAgent = UserAgent ?? http.UserAgent;
-
-            http.UserAgent = userAgent.HasValue()
-                ? userAgent
-                : "RestSharp/" + Version;
-
-            var timeout = request.Timeout != 0
-                ? request.Timeout
-                : Timeout;
-
-            if (timeout != 0)
-                http.Timeout = timeout;
-
-            var readWriteTimeout = request.ReadWriteTimeout != 0
-                ? request.ReadWriteTimeout
-                : ReadWriteTimeout;
-
-            if (readWriteTimeout != 0)
-                http.ReadWriteTimeout = readWriteTimeout;
-
-            http.FollowRedirects = FollowRedirects;
-
-            if (ClientCertificates != null)
-                http.ClientCertificates = ClientCertificates;
-
-            http.MaxRedirects = MaxRedirects;
-            http.CachePolicy  = CachePolicy;
-            http.Pipelined    = Pipelined;
-
-            if (request.Credentials != null)
-                http.Credentials = request.Credentials;
-
-            if (!IsNullOrEmpty(ConnectionGroupName))
-                http.ConnectionGroupName = ConnectionGroupName;
-
-            http.Headers = requestParameters
-                .Where(p => p.Type == ParameterType.HttpHeader)
-                .Select(p => new HttpHeader(p.Name, p.Value))
-                .ToList();
-
-            http.Cookies = requestParameters
-                .Where(p => p.Type == ParameterType.Cookie)
-                .Select(p => new HttpCookie {Name = p.Name, Value = p.Value?.ToString() ?? ""})
-                .ToList();
-
-            http.Parameters = requestParameters
-                .Where(p => p.Type == ParameterType.GetOrPost)
-                .Select(p => new HttpParameter(p.Name, p.Value))
-                .ToList();
-
-            http.Files = request.Files.Select(
-                    file => new HttpFile
-                    {
-                        Name          = file.Name,
-                        ContentType   = file.ContentType,
-                        Writer        = file.Writer,
-                        FileName      = file.FileName,
-                        ContentLength = file.ContentLength
-                    }
-                )
-                .ToList();
-
-            if (request.Body != null)
-                http.AddBody(request.Body);
-
-            http.AllowedDecompressionMethods = request.AllowedDecompressionMethods;
-
-            var proxy = Proxy ?? WebRequest.DefaultWebProxy;
-
-            try
-            {
-                proxy ??= WebRequest.GetSystemWebProxy();
-            }
-            catch (PlatformNotSupportedException)
-            {
-                // Ignore platform unsupported proxy detection
-            }
-
-            http.Proxy = proxy;
-
-            http.RemoteCertificateValidationCallback = RemoteCertificateValidationCallback;
-
-            return http;
-        }
-
-        IRestResponse<T> Deserialize<T>(IRestRequest request, IRestResponse raw)
-        {
-            IRestResponse<T> response = new RestResponse<T>();
-
-            try
-            {
-                request.OnBeforeDeserialization?.Invoke(raw);
-
-                response = raw.ToAsyncResponse<T>();
-
-                // Only attempt to deserialize if the request has not errored due
-                // to a transport or framework exception.  HTTP errors should attempt to 
-                // be deserialized 
-                if (response.ErrorException == null)
-                {
-                    var func = GetHandler(raw.ContentType);
-                    var handler = func?.Invoke();
-
-                    // Only continue if there is a handler defined else there is no way to deserialize the data.
-                    // This can happen when a request returns for example a 404 page instead of the requested JSON/XML resource
-                    if (handler is IXmlDeserializer xml)
-                    {
-                        if (request.DateFormat.IsNotEmpty())
-                            xml.DateFormat = request.DateFormat;
-
-                        if (request.XmlNamespace.IsNotEmpty())
-                            xml.Namespace = request.XmlNamespace;
-                    }
-
-                    if (handler is IWithRootElement deserializer && !request.RootElement.IsEmpty())
-                        deserializer.RootElement = request.RootElement;
-
-                    if (handler != null)
-                        response.Data = handler.Deserialize<T>(raw);
-                }
-            }
-            catch (Exception ex)
-            {
-                if (ThrowOnAnyError) throw;
-
-                if (FailOnDeserializationError || ThrowOnDeserializationError)
-                    response.ResponseStatus = ResponseStatus.Error;
-
-                response.ErrorMessage   = ex.Message;
-                response.ErrorException = ex;
-
-                if (ThrowOnDeserializationError)
-                    throw new DeserializationException(response, ex);
-            }
-
-            response.Request = request;
-
-            return response;
-        }
-
-        static bool IsWildcardStructuredSuffixSyntax(string contentType)
-        {
-            var i = 0;
-
-            // Avoid most unnecessary uses of RegEx by checking for necessary characters explicitly first
-            if (contentType[i++] != '*')
-                return false;
-
-            if (contentType[i++] != '+')
-                return false;
-
-            // If no more characters to check, exit now
-            if (i == contentType.Length)
-                return false;
-
-            // At this point it is probably using a wildcard structured syntax suffix, but let's confirm.
-            return StructuredSyntaxSuffixWildcardRegex.IsMatch(contentType);
-        }
-
-        class UrlSegmentParamsValues
-        {
-            public UrlSegmentParamsValues(Uri builderUri, string assembled)
-            {
-                Uri      = builderUri;
-                Resource = assembled;
-            }
-
-            public Uri Uri { get; }
-            public string Resource { get; }
-        }
+    }
+
+    public void Dispose() {
+        Dispose(true);
+        GC.SuppressFinalize(this);
     }
 }
