@@ -1,4 +1,4 @@
-﻿//   Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+﻿//   Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -12,86 +12,79 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License. 
 
-using System.IO;
 using System.Text;
 using System.Xml.Serialization;
-using RestSharp.Serialization.Xml;
 
-namespace RestSharp.Serializers
-{
+namespace RestSharp.Serializers.Xml;
+
+/// <summary>
+/// Wrapper for System.Xml.Serialization.XmlSerializer.
+/// </summary>
+public class DotNetXmlSerializer : IXmlSerializer {
     /// <summary>
-    /// Wrapper for System.Xml.Serialization.XmlSerializer.
+    /// Default constructor, does not specify namespace
     /// </summary>
-    public class DotNetXmlSerializer : IXmlSerializer
-    {
-        /// <summary>
-        /// Default constructor, does not specify namespace
-        /// </summary>
-        public DotNetXmlSerializer()
-        {
-            ContentType = Serialization.ContentType.Xml;
-            Encoding    = Encoding.UTF8;
-        }
+    public DotNetXmlSerializer() {
+        ContentType = Serializers.ContentType.Xml;
+        Encoding    = Encoding.UTF8;
+    }
 
-        /// <inheritdoc />
-        /// <summary>
-        /// Specify the namespaced to be used when serializing
-        /// </summary>
-        /// <param name="namespace">XML namespace</param>
-        public DotNetXmlSerializer(string @namespace)
-            : this()
-            => Namespace = @namespace;
+    /// <inheritdoc />
+    /// <summary>
+    /// Specify the namespaced to be used when serializing
+    /// </summary>
+    /// <param name="namespace">XML namespace</param>
+    public DotNetXmlSerializer(string @namespace) : this() => Namespace = @namespace;
 
-        /// <summary>
-        /// Encoding for serialized content
-        /// </summary>
-        public Encoding Encoding { get; set; }
+    /// <summary>
+    /// Encoding for serialized content
+    /// </summary>
+    public Encoding Encoding { get; set; }
 
-        /// <summary>
-        /// Serialize the object as XML
-        /// </summary>
-        /// <param name="obj">Object to serialize</param>
-        /// <returns>XML as string</returns>
-        public string Serialize(object obj)
-        {
-            var ns = new XmlSerializerNamespaces();
+    /// <summary>
+    /// Serialize the object as XML
+    /// </summary>
+    /// <param name="obj">Object to serialize</param>
+    /// <returns>XML as string</returns>
+    public string Serialize(object obj) {
+        var ns = new XmlSerializerNamespaces();
 
-            ns.Add(string.Empty, Namespace);
+        ns.Add(string.Empty, Namespace);
 
-            var serializer = new System.Xml.Serialization.XmlSerializer(obj.GetType());
-            var writer     = new EncodingStringWriter(Encoding);
+        var root = RootElement == null ? null : new XmlRootAttribute(RootElement);
 
-            serializer.Serialize(writer, obj, ns);
+        var serializer = new XmlSerializer(obj.GetType(), root);
+        var writer     = new EncodingStringWriter(Encoding);
 
-            return writer.ToString();
-        }
+        serializer.Serialize(writer, obj, ns);
 
-        /// <summary>
-        /// Name of the root element to use when serializing
-        /// </summary>
-        public string RootElement { get; set; }
+        return writer.ToString();
+    }
 
-        /// <summary>
-        /// XML namespace to use when serializing
-        /// </summary>
-        public string Namespace { get; set; }
+    /// <summary>
+    /// Name of the root element to use when serializing
+    /// </summary>
+    public string? RootElement { get; set; }
 
-        /// <summary>
-        /// Format string to use when serializing dates
-        /// </summary>
-        public string DateFormat { get; set; }
+    /// <summary>
+    /// XML namespace to use when serializing
+    /// </summary>
+    public string? Namespace { get; set; }
 
-        /// <summary>
-        /// Content type for serialized content
-        /// </summary>
-        public string ContentType { get; set; }
+    /// <summary>
+    /// Format string to use when serializing dates
+    /// </summary>
+    public string? DateFormat { get; set; }
 
-        class EncodingStringWriter : StringWriter
-        {
-            // Need to subclass StringWriter in order to override Encoding
-            public EncodingStringWriter(Encoding encoding) => Encoding = encoding;
+    /// <summary>
+    /// Content type for serialized content
+    /// </summary>
+    public string ContentType { get; set; }
 
-            public override Encoding Encoding { get; }
-        }
+    class EncodingStringWriter : StringWriter {
+        // Need to subclass StringWriter in order to override Encoding
+        public EncodingStringWriter(Encoding encoding) => Encoding = encoding;
+
+        public override Encoding Encoding { get; }
     }
 }
