@@ -1,4 +1,4 @@
-//   Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+//   Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -13,31 +13,27 @@
 //   limitations under the License. 
 
 using System.Text;
-using RestSharp.Deserializers;
-using RestSharp.Serializers;
 
-namespace RestSharp.Serialization.Xml
-{
-    public static class DotNetXmlSerializerClientExtensions
-    {
-        public static IRestClient UseDotNetXmlSerializer(
-            this IRestClient restClient,
-            string xmlNamespace = null,
-            Encoding encoding = null
-        )
-        {
-            var xmlSerializer                                 = new DotNetXmlSerializer();
-            if (xmlNamespace != null) xmlSerializer.Namespace = xmlNamespace;
-            if (encoding     != null) xmlSerializer.Encoding  = encoding;
+namespace RestSharp.Serializers.Xml; 
 
-            var xmlDeserializer                            = new DotNetXmlDeserializer();
-            if (encoding != null) xmlDeserializer.Encoding = encoding;
+[PublicAPI]
+public static class DotNetXmlSerializerClientExtensions {
+    public static RestClient UseDotNetXmlSerializer(
+        this RestClient restClient,
+        string?          xmlNamespace = null,
+        Encoding?        encoding     = null
+    ) {
+        var xmlSerializer                                 = new DotNetXmlSerializer();
+        if (xmlNamespace != null) xmlSerializer.Namespace = xmlNamespace;
+        if (encoding != null) xmlSerializer.Encoding      = encoding;
 
-            var serializer = new XmlRestSerializer()
-                .WithXmlSerializer(xmlSerializer)
-                .WithXmlDeserializer(xmlDeserializer);
+        var xmlDeserializer                            = new DotNetXmlDeserializer();
+        if (encoding != null) xmlDeserializer.Encoding = encoding;
 
-            return restClient.UseSerializer(() => serializer);
-        }
+        var serializer = new XmlRestSerializer()
+            .WithXmlSerializer(xmlSerializer)
+            .WithXmlDeserializer(xmlDeserializer);
+
+        return restClient.UseSerializer(() => serializer);
     }
 }
