@@ -1,9 +1,4 @@
-﻿using BenchmarkDotNet.Running;
+﻿using System.Reflection;
+using BenchmarkDotNet.Running;
 
-namespace RestSharp.Benchmarks
-{
-    class Program
-    {
-        public static void Main(string[] args) => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
-    }
-}
+BenchmarkSwitcher.FromAssembly(Assembly.GetExecutingAssembly()).Run(args);
