@@ -1,33 +1,13 @@
-//  Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-// http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// 
+namespace RestSharp.Benchmarks.Serializers;
 
-using System.Collections.Generic;
+public class TestClass {
+    public string         SimpleString { get; set; }
+    public int            SimpleInt    { get; set; }
+    public List<Subclass> List         { get; set; }
+    public Subclass       Sub          { get; set; }
 
-namespace RestSharp.Benchmarks.Serializers
-{
-    public class TestClass
-    {
-        public string SimpleString { get; set; }
-        public int SimpleInt { get; set; }
-        public List<Subclass> List { get; set; }
-        public Subclass Sub { get; set; }
-
-        public class Subclass
-        {
-            public string Thing { get; set; }
-            public int AnotherThing { get; set; }
-        }
+    public class Subclass {
+        public string Thing        { get; set; }
+        public int    AnotherThing { get; set; }
     }
 }
