@@ -4,57 +4,53 @@ module.exports = {
     plugins: ["@vuepress/active-header-links"],
     themeConfig: {
         logo: "/restsharp.png",
-        nav: [
-            {text: "Get help", link: "/get-help/"},
-            {text: "Gitter", link: "https://gitter.im/RestSharp/RestSharp"},
+        navbar: [
+            {text: "Migration to v107", link: "/v107/"},
+            {text: "Documentation", link: "/intro.html"},
+            {text: "Get help", link: "/support/"},
             {text: "NuGet", link: "https://nuget.org/packages/RestSharp"}
         ],
         sidebarDepth: 2,
-        sidebar: [
-            {
-                title: "Getting Started",
-                path: "/getting-started/",
-                collapsable: false,
-            },
-            {
-                title: "Using RestSharp",
-                path: "/usage/",
-                collapsable: false,
-                children: [
-                    "/usage/serialization",
-                    "/usage/files",
-                    "/usage/authenticators",
-                    "/usage/parameters",
-                    "/usage/exceptions"
-                ]
-            },
-            {
-                title: "Got stuck?",
-                path: "/get-help/",
-                collapsable: false,
-                children: [
-                    "/get-help/faq"
-                ]
-            },
-            {
-                title: "Reference",
-                path: "/api/",
-                collapsable: true,
-                children: [
-                    "/api/RestSharp",
-                    "/api/RestSharp.Serializers.NewtonsoftJson",
-                    "/api/RestSharp.Serializers.SystemTextJson",
-                    "/api/RestSharp.Serializers.Utf8Json",
-                ]
-            }
-        ],
+        sidebar: {
+            "/": [
+                {
+                    text: "",
+                    header: "RestSharp",
+                    children: [
+                        "intro.md",
+                        "usage.md",
+                        "serialization.md",
+                        "authenticators.md",
+                        "error-handling.md"
+                    ]
+                }
+            ],
+            "/v107/": [
+                {
+                    text: "",
+                    header: "Migration to v107",
+                    children: [
+                        "/v107/README.md"
+                    ]
+                }
+            ],
+            "/support/": [
+                {
+                    text: "",
+                    header: "Get help",
+                    children: [
+                        "/support/README.md"
+                    ]
+                }
+            ]
+        },
         searchPlaceholder: "Search...",
         lastUpdated: "Last Updated",
         repo: "restsharp/RestSharp",
 
         docsRepo: "restsharp/RestSharp",
         docsDir: "docs",
-        docsBranch: "master",
+        docsBranch: "dev",
         editLinks: true,
         editLinkText: "Help us by improving this page!"
     }
