@@ -1,71 +1,83 @@
-using System;
-using System.IO;
-using Newtonsoft.Json;
+//   Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License. 
+
 using Newtonsoft.Json.Serialization;
-using RestSharp.Serialization;
 
-namespace RestSharp.Serializers.NewtonsoftJson
-{
-    public class JsonNetSerializer : IRestSerializer
-    {
-        /// <summary>
-        /// Default serialization settings:
-        /// - Camel-case contract resolver
-        /// - Type name handling set to none
-        /// - Null values ignored
-        /// - Non-indented formatting
-        /// - Allow using non-public constructors
-        /// </summary>
-        public static readonly JsonSerializerSettings DefaultSettings = new JsonSerializerSettings
-        {
-            ContractResolver     = new CamelCasePropertyNamesContractResolver(),
-            DefaultValueHandling = DefaultValueHandling.Include,
-            TypeNameHandling     = TypeNameHandling.None,
-            NullValueHandling    = NullValueHandling.Ignore,
-            Formatting           = Formatting.None,
-            ConstructorHandling  = ConstructorHandling.AllowNonPublicDefaultConstructor
-        };
+namespace RestSharp.Serializers.NewtonsoftJson;
 
-        [ThreadStatic] static WriterBuffer tWriterBuffer;
+public class JsonNetSerializer : IRestSerializer, ISerializer, IDeserializer {
+    /// <summary>
+    /// Default serialization settings:
+    /// - Camel-case contract resolver
+    /// - Type name handling set to none
+    /// - Null values ignored
+    /// - Non-indented formatting
+    /// - Allow using non-public constructors
+    /// </summary>
+    public static readonly JsonSerializerSettings DefaultSettings = new() {
+        ContractResolver     = new CamelCasePropertyNamesContractResolver(),
+        DefaultValueHandling = DefaultValueHandling.Include,
+        TypeNameHandling     = TypeNameHandling.None,
+        NullValueHandling    = NullValueHandling.Ignore,
+        Formatting           = Formatting.None,
+        ConstructorHandling  = ConstructorHandling.AllowNonPublicDefaultConstructor
+    };
 
-        readonly JsonSerializer _serializer;
+    [ThreadStatic] static WriterBuffer? _writerBuffer;
 
-        /// <summary>
-        /// Create the new serializer that uses Json.Net with default settings
-        /// </summary>
-        public JsonNetSerializer() => _serializer = JsonSerializer.Create(DefaultSettings);
+    readonly JsonSerializer _serializer;
 
-        /// <summary>
-        /// Create the new serializer that uses Json.Net with custom settings
-        /// </summary>
-        /// <param name="settings">Json.Net serializer settings</param>
-        public JsonNetSerializer(JsonSerializerSettings settings) => _serializer = JsonSerializer.Create(settings);
+    /// <summary>
+    /// Create the new serializer that uses Json.Net with default settings
+    /// </summary>
+    public JsonNetSerializer() => _serializer = JsonSerializer.Create(DefaultSettings);
 
-        public string Serialize(object obj)
-        {
-            using var writerBuffer = tWriterBuffer ??= new WriterBuffer(_serializer);
+    /// <summary>
+    /// Create the new serializer that uses Json.Net with custom settings
+    /// </summary>
+    /// <param name="settings">Json.Net serializer settings</param>
+    public JsonNetSerializer(JsonSerializerSettings settings) => _serializer = JsonSerializer.Create(settings);
 
-            _serializer.Serialize(writerBuffer.GetJsonTextWriter(), obj, obj.GetType());
+    public string? Serialize(object? obj) {
+        if (obj == null) return null;
 
-            return writerBuffer.GetStringWriter().ToString();
-        }
+        using var writerBuffer = _writerBuffer ??= new WriterBuffer(_serializer);
 
-        public string Serialize(Parameter bodyParameter) => Serialize(bodyParameter.Value);
+        _serializer.Serialize(writerBuffer.GetJsonTextWriter(), obj, obj.GetType());
 
-        public T Deserialize<T>(IRestResponse response)
-        {
-            using var reader = new JsonTextReader(new StringReader(response.Content)) {CloseInput = true};
-
-            return _serializer.Deserialize<T>(reader);
-        }
-
-        public string[] SupportedContentTypes { get; } =
-        {
-            "application/json", "text/json", "text/x-json", "text/javascript", "*+json"
-        };
-
-        public string ContentType { get; set; } = "application/json";
-
-        public DataFormat DataFormat { get; } = DataFormat.Json;
+        return writerBuffer.GetStringWriter().ToString();
     }
+
+    public string? Serialize(Parameter bodyParameter) => Serialize(bodyParameter.Value);
+
+    public T? Deserialize<T>(RestResponse response) {
+        if (response.Content == null)
+            throw new DeserializationException(response, new InvalidOperationException("Response content is null"));
+
+        using var reader = new JsonTextReader(new StringReader(response.Content)) { CloseInput = true };
+
+        return _serializer.Deserialize<T>(reader);
+    }
+
+    public ISerializer   Serializer   => this;
+    public IDeserializer Deserializer => this;
+
+    public string[] AcceptedContentTypes => Serializers.ContentType.JsonAccept;
+
+    public string ContentType { get; set; } = "application/json";
+
+    public SupportsContentType SupportsContentType => contentType => contentType.Contains("json");
+
+    public DataFormat DataFormat => DataFormat.Json;
 }
