@@ -1,4 +1,4 @@
-﻿//   Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+﻿//   Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -12,452 +12,364 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License. 
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Threading;
-using System.Threading.Tasks;
-using RestSharp.Serialization;
+using System.Runtime.CompilerServices;
+using RestSharp.Extensions;
+using RestSharp.Serializers;
 
-namespace RestSharp
-{
-    public static partial class RestClientExtensions
-    {
-        /// <summary>
-        /// Execute the request and returns a response with the dynamic object as Data
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <returns></returns>
-        public static IRestResponse<dynamic> ExecuteDynamic(this IRestClient client, IRestRequest request)
-        {
-            var     response = client.Execute<dynamic>(request);
-            var     generic  = (RestResponse<dynamic>) response;
-            dynamic content  = client.Deserialize<object>(response);
+namespace RestSharp;
 
-            generic.Data = content;
+[PublicAPI]
+public static partial class RestClientExtensions {
+    /// <summary>
+    /// Executes a GET-style request asynchronously, authenticating if needed.
+    /// The response content then gets deserialized to T.
+    /// </summary>
+    /// <typeparam name="T">Target deserialization type</typeparam>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    /// <returns>Deserialized response content</returns>
+    public static Task<RestResponse<T>> ExecuteGetAsync<T>(this RestClient client, RestRequest request, CancellationToken cancellationToken = default)
+        => client.ExecuteAsync<T>(request, Method.Get, cancellationToken);
 
-            return generic;
+    /// <summary>
+    /// Executes a GET-style asynchronously, authenticating if needed
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    public static Task<RestResponse> ExecuteGetAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default)
+        => client.ExecuteAsync(request, Method.Get, cancellationToken);
+
+    /// <summary>
+    /// Executes a POST-style request asynchronously, authenticating if needed.
+    /// The response content then gets deserialized to T.
+    /// </summary>
+    /// <typeparam name="T">Target deserialization type</typeparam>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="cancellationToken">The cancellation token</param>
+    /// <returns>Deserialized response content</returns>
+    public static Task<RestResponse<T>> ExecutePostAsync<T>(
+        this RestClient   client,
+        RestRequest       request,
+        CancellationToken cancellationToken = default
+    )
+        => client.ExecuteAsync<T>(request, Method.Post, cancellationToken);
+
+    /// <summary>
+    /// Executes a POST-style asynchronously, authenticating if needed
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    public static Task<RestResponse> ExecutePostAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default)
+        => client.ExecuteAsync(request, Method.Post, cancellationToken);
+
+    /// <summary>
+    /// Executes a PUT-style request asynchronously, authenticating if needed.
+    /// The response content then gets deserialized to T.
+    /// </summary>
+    /// <typeparam name="T">Target deserialization type</typeparam>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="cancellationToken">The cancellation token</param>
+    /// <returns>Deserialized response content</returns>
+    public static Task<RestResponse<T>> ExecutePutAsync<T>(
+        this RestClient   client,
+        RestRequest       request,
+        CancellationToken cancellationToken = default
+    )
+        => client.ExecuteAsync<T>(request, Method.Put, cancellationToken);
+
+    /// <summary>
+    /// Executes a PUP-style asynchronously, authenticating if needed
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    public static Task<RestResponse> ExecutePutAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default)
+        => client.ExecuteAsync(request, Method.Put, cancellationToken);
+
+    /// <summary>
+    /// Executes the request asynchronously, authenticating if needed
+    /// </summary>
+    /// <typeparam name="T">Target deserialization type</typeparam>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    public static async Task<RestResponse<T>> ExecuteAsync<T>(
+        this RestClient   client,
+        RestRequest       request,
+        CancellationToken cancellationToken = default
+    ) {
+        if (request == null)
+            throw new ArgumentNullException(nameof(request));
+
+        var response = await client.ExecuteAsync(request, cancellationToken).ConfigureAwait(false);
+        return client.Deserialize<T>(request, response);
+    }
+
+    /// <summary>
+    /// Executes the request asynchronously, authenticating if needed
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="httpMethod">Override the request method</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    public static Task<RestResponse> ExecuteAsync(
+        this RestClient   client,
+        RestRequest       request,
+        Method            httpMethod,
+        CancellationToken cancellationToken = default
+    ) {
+        Ensure.NotNull(request, nameof(request));
+
+        request.Method = httpMethod;
+        return client.ExecuteAsync(request, cancellationToken);
+    }
+
+    /// <summary>
+    /// Executes the request asynchronously, authenticating if needed
+    /// </summary>
+    /// <typeparam name="T">Target deserialization type</typeparam>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="httpMethod">Override the request method</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    public static Task<RestResponse<T>> ExecuteAsync<T>(
+        this RestClient   client,
+        RestRequest       request,
+        Method            httpMethod,
+        CancellationToken cancellationToken = default
+    ) {
+        Ensure.NotNull(request, nameof(request));
+
+        request.Method = httpMethod;
+        return client.ExecuteAsync<T>(request, cancellationToken);
+    }
+
+    /// <summary>
+    /// Execute the request using GET HTTP method. Exception will be thrown if the request does not succeed.
+    /// The response data is deserialized to the Data property of the returned response object.
+    /// </summary>
+    /// <param name="client">RestClient instance</param>
+    /// <param name="request">The request</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    /// <typeparam name="T">Expected result type</typeparam>
+    /// <returns></returns>
+    public static async Task<T?> GetAsync<T>(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteGetAsync<T>(request, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError().Data;
+    }
+    
+    public static async Task<RestResponse> GetAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteGetAsync(request, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError();
+    }
+
+    /// <summary>
+    /// Execute the request using POST HTTP method. Exception will be thrown if the request does not succeed.
+    /// The response data is deserialized to the Data property of the returned response object.
+    /// </summary>
+    /// <param name="client">RestClient instance</param>
+    /// <param name="request">The request</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    /// <typeparam name="T">Expected result type</typeparam>
+    /// <returns></returns>
+    public static async Task<T?> PostAsync<T>(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecutePostAsync<T>(request, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError().Data;
+    }
+
+    public static RestResponse Post(this RestClient client, RestRequest request)
+        => AsyncHelpers.RunSync(() => client.PostAsync(request));
+
+    public static async Task<RestResponse> PostAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecutePostAsync(request, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError();
+    }
+
+    /// <summary>
+    /// Execute the request using PUT HTTP method. Exception will be thrown if the request does not succeed.
+    /// The response data is deserialized to the Data property of the returned response object.
+    /// </summary>
+    /// <param name="client">RestClient instance</param>
+    /// <param name="request">The request</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    /// <typeparam name="T">Expected result type</typeparam>
+    /// <returns></returns>
+    public static async Task<T?> PutAsync<T>(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync<T>(request, Method.Put, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError().Data;
+    }
+
+    public static async Task<RestResponse> PutAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync(request, Method.Put, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError();
+    }
+
+    /// <summary>
+    /// Execute the request using HEAD HTTP method. Exception will be thrown if the request does not succeed.
+    /// The response data is deserialized to the Data property of the returned response object.
+    /// </summary>
+    /// <param name="client">RestClient instance</param>
+    /// <param name="request">The request</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    /// <typeparam name="T">Expected result type</typeparam>
+    /// <returns></returns>
+    public static async Task<T?> HeadAsync<T>(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync<T>(request, Method.Head, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError().Data;
+    }
+
+    public static async Task<RestResponse> HeadAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync(request, Method.Head, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError();
+    }
+
+    /// <summary>
+    /// Execute the request using OPTIONS HTTP method. Exception will be thrown if the request does not succeed.
+    /// The response data is deserialized to the Data property of the returned response object.
+    /// </summary>
+    /// <param name="client">RestClient instance</param>
+    /// <param name="request">The request</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    /// <typeparam name="T">Expected result type</typeparam>
+    /// <returns></returns>
+    public static async Task<T?> OptionsAsync<T>(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync<T>(request, Method.Options, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError().Data;
+    }
+
+    public static async Task<RestResponse> OptionsAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync(request, Method.Options, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError();
+    }
+
+    /// <summary>
+    /// Execute the request using PATCH HTTP method. Exception will be thrown if the request does not succeed.
+    /// The response data is deserialized to the Data property of the returned response object.
+    /// </summary>
+    /// <param name="client">RestClient instance</param>
+    /// <param name="request">The request</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    /// <typeparam name="T">Expected result type</typeparam>
+    /// <returns></returns>
+    public static async Task<T?> PatchAsync<T>(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync<T>(request, Method.Patch, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError().Data;
+    }
+
+    public static async Task<RestResponse> PatchAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync(request, Method.Patch, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError();
+    }
+
+    /// <summary>
+    /// Execute the request using DELETE HTTP method. Exception will be thrown if the request does not succeed.
+    /// The response data is deserialized to the Data property of the returned response object.
+    /// </summary>
+    /// <param name="client">RestClient instance</param>
+    /// <param name="request">The request</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    /// <typeparam name="T">Expected result type</typeparam>
+    /// <returns></returns>
+    public static async Task<T?> DeleteAsync<T>(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync<T>(request, Method.Delete, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError().Data;
+    }
+
+    public static async Task<RestResponse> DeleteAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+        var response = await client.ExecuteAsync(request, Method.Delete, cancellationToken).ConfigureAwait(false);
+        return response.ThrowIfError();
+    }
+
+    /// <summary>
+    /// A specialized method to download files.
+    /// </summary>
+    /// <param name="client">RestClient instance</param>
+    /// <param name="request">Pre-configured request instance.</param>
+    /// <param name="cancellationToken"></param>
+    /// <returns>The downloaded file.</returns>
+    [PublicAPI]
+    public static async Task<byte[]?> DownloadDataAsync(this RestClient client, RestRequest request, CancellationToken cancellationToken = default) {
+#if NETSTANDARD
+        using var stream = await client.DownloadStreamAsync(request, cancellationToken).ConfigureAwait(false);
+#else
+        await using var stream = await client.DownloadStreamAsync(request, cancellationToken).ConfigureAwait(false);
+#endif
+        return stream == null ? null : await stream.ReadAsBytes(cancellationToken).ConfigureAwait(false);
+    }
+
+    /// <summary>
+    /// Reads a stream returned by the specified endpoint, deserializes each line to JSON and returns each object asynchronously.
+    /// It is required for each JSON object to be returned in a single line.
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="resource"></param>
+    /// <param name="cancellationToken"></param>
+    /// <typeparam name="T"></typeparam>
+    /// <returns></returns>
+    [PublicAPI]
+    public static async IAsyncEnumerable<T> StreamJsonAsync<T>(
+        this RestClient                            client,
+        string                                     resource,
+        [EnumeratorCancellation] CancellationToken cancellationToken
+    ) {
+        var request = new RestRequest(resource);
+
+#if NETSTANDARD
+        using var stream = await client.DownloadStreamAsync(request, cancellationToken).ConfigureAwait(false);
+#else
+        await using var stream = await client.DownloadStreamAsync(request, cancellationToken).ConfigureAwait(false);
+#endif
+        if (stream == null) yield break;
+
+        var serializer = client.Serializers[DataFormat.Json].GetSerializer();
+
+        using var reader = new StreamReader(stream);
+
+        while (!reader.EndOfStream && !cancellationToken.IsCancellationRequested) {
+            var line = await reader.ReadLineAsync().ConfigureAwait(false);
+            if (string.IsNullOrWhiteSpace(line)) continue;
+
+            var response = new RestResponse { Content = line };
+            yield return serializer.Deserializer.Deserialize<T>(response)!;
         }
+    }
 
-        /// <summary>
-        /// Execute the request using GET HTTP method. Exception will be thrown if the request does not succeed.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static async Task<T> GetAsync<T>(this IRestClient client, IRestRequest request, CancellationToken cancellationToken = default)
-        {
-            var response = await client.ExecuteGetAsync<T>(request, cancellationToken);
-            ThrowIfError(response);
-            return response.Data;
-        }
+    /// <summary>
+    /// Sets the <see cref="RestClient"/> to only use JSON
+    /// </summary>
+    /// <param name="client">Client instance to work with</param>
+    /// <returns>Reference to the client instance</returns>
+    public static RestClient UseJson(this RestClient client) {
+        client.Serializers.Remove(DataFormat.Xml);
+        client.AssignAcceptedContentTypes();
+        return client;
+    }
 
-        /// <summary>
-        /// Execute the request using POST HTTP method. Exception will be thrown if the request does not succeed.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static async Task<T> PostAsync<T>(this IRestClient client, IRestRequest request, CancellationToken cancellationToken = default)
-        {
-            var response = await client.ExecutePostAsync<T>(request, cancellationToken);
-            ThrowIfError(response);
-            return response.Data;
-        }
+    /// <summary>
+    /// Sets the <see cref="RestClient"/> to only use XML
+    /// </summary>
+    /// <param name="client">Client instance to work with</param>
+    /// <returns>Reference to the client instance</returns>
+    public static RestClient UseXml(this RestClient client) {
+        client.Serializers.Remove(DataFormat.Json);
+        client.AssignAcceptedContentTypes();
+        return client;
+    }
 
-        /// <summary>
-        /// Execute the request using PUT HTTP method. Exception will be thrown if the request does not succeed.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static async Task<T> PutAsync<T>(this IRestClient client, IRestRequest request, CancellationToken cancellationToken = default)
-        {
-            var response = await client.ExecuteAsync<T>(request, Method.PUT, cancellationToken);
-            ThrowIfError(response);
-            return response.Data;
-        }
-
-        /// <summary>
-        /// Execute the request using HEAD HTTP method. Exception will be thrown if the request does not succeed.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static async Task<T> HeadAsync<T>(this IRestClient client, IRestRequest request, CancellationToken cancellationToken = default)
-        {
-            var response = await client.ExecuteAsync<T>(request, Method.HEAD, cancellationToken);
-            ThrowIfError(response);
-            return response.Data;
-        }
-
-        /// <summary>
-        /// Execute the request using OPTIONS HTTP method. Exception will be thrown if the request does not succeed.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static async Task<T> OptionsAsync<T>(this IRestClient client, IRestRequest request, CancellationToken cancellationToken = default)
-        {
-            var response = await client.ExecuteAsync<T>(request, Method.OPTIONS, cancellationToken);
-            ThrowIfError(response);
-            return response.Data;
-        }
-
-        /// <summary>
-        /// Execute the request using PATCH HTTP method. Exception will be thrown if the request does not succeed.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static async Task<T> PatchAsync<T>(this IRestClient client, IRestRequest request, CancellationToken cancellationToken = default)
-        {
-            var response = await client.ExecuteAsync<T>(request, Method.PATCH, cancellationToken);
-            ThrowIfError(response);
-            return response.Data;
-        }
-
-        /// <summary>
-        /// Execute the request using DELETE HTTP method. Exception will be thrown if the request does not succeed.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static async Task<T> DeleteAsync<T>(this IRestClient client, IRestRequest request, CancellationToken cancellationToken = default)
-        {
-            var response = await client.ExecuteAsync<T>(request, Method.DELETE, cancellationToken);
-            ThrowIfError(response);
-            return response.Data;
-        }
-
-        /// <summary>
-        /// Execute the request using GET HTTP method.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static IRestResponse<T> Get<T>(this IRestClient client, IRestRequest request)
-            => client.Execute<T>(request, Method.GET);
-
-        /// <summary>
-        /// Execute the request using POST HTTP method.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static IRestResponse<T> Post<T>(this IRestClient client, IRestRequest request)
-            => client.Execute<T>(request, Method.POST);
-
-        /// <summary>
-        /// Execute the request using PUT HTTP method.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static IRestResponse<T> Put<T>(this IRestClient client, IRestRequest request)
-            => client.Execute<T>(request, Method.PUT);
-
-        /// <summary>
-        /// Execute the request using HEAD HTTP method.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static IRestResponse<T> Head<T>(this IRestClient client, IRestRequest request)
-            => client.Execute<T>(request, Method.HEAD);
-
-        /// <summary>
-        /// Execute the request using OPTIONS HTTP method.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static IRestResponse<T> Options<T>(this IRestClient client, IRestRequest request)
-            => client.Execute<T>(request, Method.OPTIONS);
-
-        /// <summary>
-        /// Execute the request using PATCH HTTP method.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static IRestResponse<T> Patch<T>(this IRestClient client, IRestRequest request)
-            => client.Execute<T>(request, Method.PATCH);
-
-        /// <summary>
-        /// Execute the request using DELETE HTTP method.
-        /// The response data is deserialzied to the Data property of the returned response object.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <typeparam name="T">Expected result type</typeparam>
-        /// <returns></returns>
-        public static IRestResponse<T> Delete<T>(this IRestClient client, IRestRequest request)
-            => client.Execute<T>(request, Method.DELETE);
-
-        /// <summary>
-        /// Execute the request using GET HTTP method.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <returns></returns>
-        public static IRestResponse Get(this IRestClient client, IRestRequest request) => client.Execute(request, Method.GET);
-
-        /// <summary>
-        /// Execute the request using POST HTTP method.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <returns></returns>
-        public static IRestResponse Post(this IRestClient client, IRestRequest request) => client.Execute(request, Method.POST);
-
-        /// <summary>
-        /// Execute the request using PUT HTTP method.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <returns></returns>
-        public static IRestResponse Put(this IRestClient client, IRestRequest request) => client.Execute(request, Method.PUT);
-
-        /// <summary>
-        /// Execute the request using HEAD HTTP method.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <returns></returns>
-        public static IRestResponse Head(this IRestClient client, IRestRequest request) => client.Execute(request, Method.HEAD);
-
-        /// <summary>
-        /// Execute the request using OPTIONS HTTP method.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <returns></returns>
-        public static IRestResponse Options(this IRestClient client, IRestRequest request) => client.Execute(request, Method.OPTIONS);
-
-        /// <summary>
-        /// Execute the request using PATCH HTTP method.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <returns></returns>
-        public static IRestResponse Patch(this IRestClient client, IRestRequest request) => client.Execute(request, Method.PATCH);
-
-        /// <summary>
-        /// Execute the request using DELETE HTTP method.
-        /// </summary>
-        /// <param name="client">RestClient instance</param>
-        /// <param name="request">The request</param>
-        /// <returns></returns>
-        public static IRestResponse Delete(this IRestClient client, IRestRequest request) => client.Execute(request, Method.DELETE);
-
-        /// <summary>
-        /// Add a parameter to use on every request made with this client instance
-        /// </summary>
-        /// <param name="restClient">The IRestClient instance</param>
-        /// <param name="p">Parameter to add</param>
-        /// <returns></returns>
-        public static IRestClient AddDefaultParameter(this IRestClient restClient, Parameter p)
-        {
-            if (p.Type == ParameterType.RequestBody)
-                throw new NotSupportedException(
-                    "Cannot set request body from default headers. Use Request.AddBody() instead."
-                );
-
-            restClient.DefaultParameters.Add(p);
-
-            return restClient;
-        }
-
-        /// <summary>
-        /// Add a new or update an existing parameter to use on every request made with this client instance
-        /// </summary>
-        /// <param name="restClient"></param>
-        /// <param name="p"></param>
-        /// <returns></returns>
-        public static IRestClient AddOrUpdateDefaultParameter(this IRestClient restClient, Parameter p)
-        {
-            var existing = restClient.DefaultParameters.FirstOrDefault(x => x.Name == p.Name);
-
-            if (existing != null) restClient.DefaultParameters.Remove(existing);
-
-            restClient.DefaultParameters.Add(p);
-
-            return restClient;
-        }
-
-        /// <summary>
-        /// Removes a parameter from the default parameters that are used on every request made with this client instance
-        /// </summary>
-        /// <param name="restClient">The IRestClient instance</param>
-        /// <param name="name">The name of the parameter that needs to be removed</param>
-        /// <returns></returns>
-        public static IRestClient RemoveDefaultParameter(this IRestClient restClient, string name)
-        {
-            var parameter = restClient.DefaultParameters.SingleOrDefault(
-                p => p.Name.Equals(name, StringComparison.OrdinalIgnoreCase)
-            );
-
-            if (parameter != null) restClient.DefaultParameters.Remove(parameter);
-
-            return restClient;
-        }
-
-        /// <summary>
-        /// Adds a default HTTP parameter (QueryString for GET, DELETE, OPTIONS and HEAD; Encoded form for POST and PUT)
-        /// Used on every request made by this client instance
-        /// </summary>
-        /// <param name="restClient">The IRestClient instance</param>
-        /// <param name="name">Name of the parameter</param>
-        /// <param name="value">Value of the parameter</param>
-        /// <returns>This request</returns>
-        public static IRestClient AddDefaultParameter(this IRestClient restClient, string name, object value)
-            => restClient.AddDefaultParameter(new Parameter(name, value, ParameterType.GetOrPost));
-
-        /// <summary>
-        /// Adds a default parameter to the request. There are four types of parameters:
-        /// - GetOrPost: Either a QueryString value or encoded form value based on method
-        /// - HttpHeader: Adds the name/value pair to the HTTP request's Headers collection
-        /// - UrlSegment: Inserted into URL if there is a matching url token e.g. {AccountId}
-        /// - RequestBody: Used by AddBody() (not recommended to use directly)
-        /// Used on every request made by this client instance
-        /// </summary>
-        /// <param name="restClient">The IRestClient instance</param>
-        /// <param name="name">Name of the parameter</param>
-        /// <param name="value">Value of the parameter</param>
-        /// <param name="type">The type of parameter to add</param>
-        /// <returns>This request</returns>
-        public static IRestClient AddDefaultParameter(
-            this IRestClient restClient,
-            string name,
-            object value,
-            ParameterType type
-        )
-            => restClient.AddDefaultParameter(new Parameter(name, value, type));
-
-        /// <summary>
-        /// Adds a default header to the RestClient. Used on every request made by this client instance.
-        /// </summary>
-        /// <param name="restClient">The IRestClient instance</param>
-        /// <param name="name">Name of the header to add</param>
-        /// <param name="value">Value of the header to add</param>
-        /// <returns></returns>
-        public static IRestClient AddDefaultHeader(this IRestClient restClient, string name, string value)
-            => restClient.AddDefaultParameter(name, value, ParameterType.HttpHeader);
-
-        /// <summary>
-        /// Adds default headers to the RestClient. Used on every request made by this client instance.
-        /// </summary>
-        /// <param name="restClient">The IRestClient instance</param>
-        /// <param name="headers">Dictionary containing the Names and Values of the headers to add</param>
-        /// <returns></returns>
-        public static IRestClient AddDefaultHeaders(this IRestClient restClient, Dictionary<string, string> headers)
-        {
-            foreach (var header in headers)
-                restClient.AddOrUpdateDefaultParameter(new Parameter(header.Key, header.Value, ParameterType.HttpHeader));
-
-            return restClient;
-        }
-
-        /// <summary>
-        /// Adds a default URL segment parameter to the RestClient. Used on every request made by this client instance.
-        /// </summary>
-        /// <param name="restClient">The IRestClient instance</param>
-        /// <param name="name">Name of the segment to add</param>
-        /// <param name="value">Value of the segment to add</param>
-        /// <returns></returns>
-        public static IRestClient AddDefaultUrlSegment(this IRestClient restClient, string name, string value)
-            => restClient.AddDefaultParameter(name, value, ParameterType.UrlSegment);
-
-        /// <summary>
-        /// Adds a default URL query parameter to the RestClient. Used on every request made by this client instance.
-        /// </summary>
-        /// <param name="restClient">The IRestClient instance</param>
-        /// <param name="name">Name of the query parameter to add</param>
-        /// <param name="value">Value of the query parameter to add</param>
-        /// <returns></returns>
-        public static IRestClient AddDefaultQueryParameter(this IRestClient restClient, string name, string value)
-            => restClient.AddDefaultParameter(name, value, ParameterType.QueryString);
-
-        static void ThrowIfError(IRestResponse response)
-        {
-            var exception = response.ResponseStatus switch
-            {
-                ResponseStatus.Aborted   => new WebException("Request aborted", response.ErrorException),
-                ResponseStatus.Error     => response.ErrorException,
-                ResponseStatus.TimedOut  => new TimeoutException("Request timed out", response.ErrorException),
-                ResponseStatus.None      => null,
-                ResponseStatus.Completed => null,
-                _                        => throw response.ErrorException ?? new ArgumentOutOfRangeException()
-            };
-
-            if (exception != null)
-                throw exception;
-        }
-
-        /// <summary>
-        /// Sets the <see cref="RestClient"/> to only use JSON
-        /// </summary>
-        /// <param name="client">The client instance</param>
-        /// <returns></returns>
-        public static RestClient UseJson(this RestClient client)
-        {
-            foreach (var contentType in ContentType.XmlAccept)
-                client.RemoveHandler(contentType);
-
-            client.Serializers.Remove(DataFormat.Xml);
-
-            return client;
-        }
-
-        /// <summary>
-        /// Sets the <see cref="RestClient"/> to only use XML
-        /// </summary>
-        /// <param name="client"></param>
-        /// <returns></returns>
-        public static RestClient UseXml(this RestClient client)
-        {
-            foreach (var contentType in ContentType.JsonAccept)
-                client.RemoveHandler(contentType);
-
-            client.Serializers.Remove(DataFormat.Json);
-
-            return client;
-        }
+    /// <summary>
+    /// Sets the <see cref="RestClient"/> to only use the passed in custom serializer
+    /// </summary>
+    /// <param name="client">Client instance to work with</param>
+    /// <param name="serializerFactory">Function that returns the serializer instance</param>
+    /// <returns>Reference to the client instance</returns>
+    public static RestClient UseOnlySerializer(this RestClient client, Func<IRestSerializer> serializerFactory) {
+        client.Serializers.Clear();
+        client.UseSerializer(serializerFactory);
+        return client;
     }
 }
