@@ -1,4 +1,4 @@
-//   Copyright © 2009-2020 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
+//   Copyright © 2009-2021 John Sheehan, Andrew Young, Alexey Zimarev and RestSharp community
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -12,17 +12,17 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License. 
 
-using RestSharp.Deserializers;
-using RestSharp.Serializers;
+namespace RestSharp.Serializers; 
 
-namespace RestSharp.Serialization
-{
-    public interface IRestSerializer : ISerializer, IDeserializer
-    {
-        string[] SupportedContentTypes { get; }
+public interface IRestSerializer {
+    ISerializer   Serializer   { get; }
+    IDeserializer Deserializer { get; }
+    
+    string[] AcceptedContentTypes { get; }
 
-        DataFormat DataFormat { get; }
+    SupportsContentType SupportsContentType { get; }
 
-        string Serialize(Parameter parameter);
-    }
+    DataFormat DataFormat { get; }
+
+    string? Serialize(Parameter parameter);
 }
