@@ -1,30 +1,23 @@
-﻿using NUnit.Framework;
+﻿namespace RestSharp.Tests;
 
-namespace RestSharp.Tests
-{
-    [TestFixture]
-    public class AddRangeTests
-    {
-        [Test]
-        public void ShouldParseOutLongRangeSpecifier()
-        {
-            var        restClient = new RestClient("http://localhost");
-            var        req        = new RestRequest("bob", Method.GET);
-            const long start      = (long) int.MaxValue + 1;
-            const long end        = start               + 1;
+public class AddRangeTests {
+    [Fact]
+    public async Task ShouldParseOutLongRangeSpecifier() {
+        var        restClient = new RestClient("http://localhost");
+        var        req        = new RestRequest("bob", Method.Get);
+        const long start      = (long)int.MaxValue + 1;
+        const long end        = start + 1;
 
-            req.AddHeader("Range", $"pages={start}-{end}");
-            restClient.Execute(req);
-        }
+        req.AddHeader("Range", $"pages={start}-{end}");
+        await restClient.ExecuteAsync(req);
+    }
 
-        [Test]
-        public void ShouldParseOutRangeSpecifier()
-        {
-            var restClient = new RestClient("http://localhost");
-            var req        = new RestRequest("bob", Method.GET);
+    [Fact]
+    public async Task ShouldParseOutRangeSpecifier() {
+        var restClient = new RestClient("http://localhost");
+        var req        = new RestRequest("bob", Method.Get);
 
-            req.AddHeader("Range", "pages=1-2");
-            restClient.Execute(req);
-        }
+        req.AddHeader("Range", "pages=1-2");
+        await restClient.ExecuteAsync(req);
     }
 }
