@@ -1,82 +1,174 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using NUnit.Framework;
+namespace RestSharp.Tests; 
 
-namespace RestSharp.Tests
-{
-    public class RequestHeaderTests
-    {
-        [Test]
-        public void AddHeaders_SameCaseDuplicatesExist_ThrowsException()
-        {
-            var headers = new List<KeyValuePair<string, string>>
-            {
-                new KeyValuePair<string, string>("Accept", "application/json"),
-                new KeyValuePair<string, string>("Accept-Language", "en-us,en;q=0.5"),
-                new KeyValuePair<string, string>("Keep-Alive", "300"),
-                new KeyValuePair<string, string>("Accept", "application/json")
-            };
+public class RequestHeaderTests {
+    [Fact]
+    public void AddHeaders_SameCaseDuplicatesExist_ThrowsException() {
+        var headers = new List<KeyValuePair<string, string>> {
+            new("Accept", "application/json"),
+            new("Accept-Language", "en-us,en;q=0.5"),
+            new("Keep-Alive", "300"),
+            new("Accept", "application/json")
+        };
 
-            var request = new RestRequest();
+        var request = new RestRequest();
 
-            var exception = Assert.Throws<ArgumentException>(() => request.AddHeaders(headers));
-            Assert.AreEqual("Duplicate header names exist: ACCEPT", exception.Message);
-        }
+        var exception = Assert.Throws<ArgumentException>(() => request.AddHeaders(headers));
+        Assert.Equal("Duplicate header names exist: ACCEPT", exception.Message);
+    }
 
-        [Test]
-        public void AddHeaders_DifferentCaseDuplicatesExist_ThrowsException()
-        {
-            var headers = new List<KeyValuePair<string, string>>()
-            {
-                new KeyValuePair<string, string>("Accept", "application/json"),
-                new KeyValuePair<string, string>("Accept-Language", "en-us,en;q=0.5"),
-                new KeyValuePair<string, string>("Keep-Alive", "300"),
-                new KeyValuePair<string, string>("acCEpt", "application/json")
-            };
+    [Fact]
+    public void AddHeaders_DifferentCaseDuplicatesExist_ThrowsException() {
+        var headers = new List<KeyValuePair<string, string>> {
+            new("Accept", "application/json"),
+            new("Accept-Language", "en-us,en;q=0.5"),
+            new("Keep-Alive", "300"),
+            new("acCEpt", "application/json")
+        };
 
-            var request = new RestRequest();
+        var request = new RestRequest();
 
-            var exception = Assert.Throws<ArgumentException>(() => request.AddHeaders(headers));
-            Assert.AreEqual("Duplicate header names exist: ACCEPT", exception.Message);
-        }
+        var exception = Assert.Throws<ArgumentException>(() => request.AddHeaders(headers));
+        Assert.Equal("Duplicate header names exist: ACCEPT", exception.Message);
+    }
 
-        [Test]
-        public void AddHeaders_NoDuplicatesExist_Has3Headers()
-        {
-            var headers = new List<KeyValuePair<string, string>>()
-            {
-                new KeyValuePair<string, string>("Accept", "application/json"),
-                new KeyValuePair<string, string>("Accept-Language", "en-us,en;q=0.5"),
-                new KeyValuePair<string, string>("Keep-Alive", "300")
-            };
+    [Fact]
+    public void AddHeaders_NoDuplicatesExist_Has3Headers() {
+        var headers = new List<KeyValuePair<string, string>> {
+            new("Accept", "application/json"),
+            new("Accept-Language", "en-us,en;q=0.5"),
+            new("Keep-Alive", "300")
+        };
 
-            var request = new RestRequest();
-            request.AddHeaders(headers);
+        var request = new RestRequest();
+        request.AddHeaders(headers);
 
-            var httpParameters = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader);
+        var httpParameters = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader);
 
-            Assert.AreEqual(3, httpParameters.Count());
-        }
+        Assert.Equal(3, httpParameters.Count());
+    }
 
-        [Test]
-        public void AddHeaders_NoDuplicatesExistUsingDictionary_Has3Headers()
-        {
-            var headers = new Dictionary<string, string>()
-            {
-                { "Accept", "application/json" },
-                { "Accept-Language", "en-us,en;q=0.5" },
-                { "Keep-Alive", "300" }
-            };
+    [Fact]
+    public void AddHeaders_NoDuplicatesExistUsingDictionary_Has3Headers() {
+        var headers = new Dictionary<string, string> {
+            { "Accept", "application/json" },
+            { "Accept-Language", "en-us,en;q=0.5" },
+            { "Keep-Alive", "300" }
+        };
 
-            var request = new RestRequest();
-            request.AddHeaders(headers);
+        var request = new RestRequest();
+        request.AddHeaders(headers);
 
-            var httpParameters = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader);
+        var httpParameters = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader);
 
-            Assert.AreEqual(3, httpParameters.Count());
-        }
+        Assert.Equal(3, httpParameters.Count());
+    }
+
+    [Fact]
+    public void AddOrUpdateHeader_ShouldUpdateExistingHeader_WhenHeaderExist() {
+        // Arrange
+        var request = new RestRequest();
+        request.AddHeader("Accept", "application/xml");
+
+        // Act
+        request.AddOrUpdateHeader("Accept", "application/json");
+
+        // Assert
+        var headers = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
+
+        Assert.Equal("application/json", headers.First(parameter => parameter.Name == "Accept").Value);
+        Assert.Single(headers);
+    }
+
+    [Fact]
+    public void AddOrUpdateHeader_ShouldUpdateExistingHeader_WhenHeaderDoesNotExist() {
+        // Arrange
+        var request = new RestRequest();
+
+        // Act
+        request.AddOrUpdateHeader("Accept", "application/json");
+
+        // Assert
+        var headers = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
+
+        Assert.Equal("application/json", headers.First(parameter => parameter.Name == "Accept").Value);
+        Assert.Single(headers);
+    }
+
+    [Fact]
+    public void AddOrUpdateHeaders_ShouldAddHeaders_WhenNoneExists() {
+        // Arrange
+        var headers = new Dictionary<string, string> {
+            { "Accept", "application/json" },
+            { "Accept-Language", "en-us,en;q=0.5" },
+            { "Keep-Alive", "300" }
+        };
+
+        var request = new RestRequest();
+
+        // Act
+        request.AddOrUpdateHeaders(headers);
+
+        // Assert
+        var requestHeaders = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
+
+        Assert.Equal("application/json", requestHeaders.First(parameter => parameter.Name == "Accept").Value);
+        Assert.Equal("en-us,en;q=0.5", requestHeaders.First(parameter => parameter.Name == "Accept-Language").Value);
+        Assert.Equal("300", requestHeaders.First(parameter => parameter.Name == "Keep-Alive").Value);
+        Assert.Equal(3, requestHeaders.Length);
+    }
+
+    [Fact]
+    public void AddOrUpdateHeaders_ShouldUpdateHeaders_WhenAllExists() {
+        // Arrange
+        var headers = new Dictionary<string, string> {
+            { "Accept", "application/json" },
+            { "Keep-Alive", "300" }
+        };
+
+        var updatedHeaders = new Dictionary<string, string> {
+            { "Accept", "application/xml" },
+            { "Keep-Alive", "400" }
+        };
+
+        var request = new RestRequest();
+        request.AddHeaders(headers);
+
+        // Act
+        request.AddOrUpdateHeaders(updatedHeaders);
+
+        // Assert
+        var requestHeaders = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
+
+        Assert.Equal("application/xml", requestHeaders.First(parameter => parameter.Name == "Accept").Value);
+        Assert.Equal("400", requestHeaders.First(parameter => parameter.Name == "Keep-Alive").Value);
+        Assert.Equal(2, requestHeaders.Length);
+    }
+
+    [Fact]
+    public void AddOrUpdateHeaders_ShouldAddAndUpdateHeaders_WhenSomeExists() {
+        // Arrange
+        var headers = new Dictionary<string, string> {
+            { "Accept", "application/json" },
+            { "Keep-Alive", "300" }
+        };
+
+        var updatedHeaders = new Dictionary<string, string> {
+            { "Accept", "application/xml" },
+            { "Accept-Language", "en-us,en;q=0.5" }
+        };
+
+        var request = new RestRequest();
+        request.AddHeaders(headers);
+
+        // Act
+        request.AddOrUpdateHeaders(updatedHeaders);
+
+        // Assert
+        var requestHeaders = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
+
+        Assert.Equal("application/xml", requestHeaders.First(parameter => parameter.Name == "Accept").Value);
+        Assert.Equal("en-us,en;q=0.5", requestHeaders.First(parameter => parameter.Name == "Accept-Language").Value);
+        Assert.Equal("300", requestHeaders.First(parameter => parameter.Name == "Keep-Alive").Value);
+        Assert.Equal(3, requestHeaders.Length);
     }
 }
